repo_results/run-20240308-125808_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/whatsapp/12/LLMatic/user.py
class User:
	def __init__(self, email, password, profile_picture=None, status_message=None, privacy='public'):
		self.email = email
		self.password = password
		self.profile_picture = profile_picture
		self.status_message = status_message
		self.privacy = privacy

mock_db = {}

def sign_up(email, password, profile_picture=None, status_message=None, privacy='public'):
	if email in mock_db:
		return 'User already exists'
	else:
		mock_db[email] = User(email, password, profile_picture, status_message, privacy)
		return 'User created successfully'

def recover_password(email):
	if email in mock_db:
		new_password = 'new_password'
		mock_db[email].password = new_password
		return 'Password reset successful'
	else:
		return 'User does not exist'
